+++ CHANGELOG.md
@@ -8,9 +8,10 @@
 - chore: remove what has remained from poetry in the codebase
 - refactor(core): avoid truncating or coloring logs in log files
 - feat(web-ui): add web-ui service
-- fix(users): avoid setting user as sudoer when it performs a password reset
 - feat(web-ui): process input demands, dispatched on the bus
 - feat(keypad): ability to use key-combinations, set key combinations for screenshot, snapshot and quit
+- feat(web-ui): add `fields` in `InputDescription` with `InputFieldDescription` data structures to describe the fields of an input demand in detail
+- fix(users): avoid setting user as sudoer when it performs a password reset
 
 ## Version 1.0.0